CreAdd initial version of README.mdate README.md
Add initial version of README.md

This commit adds the initial version of the README.md file to the repository. The README includes sections detailing my background, current projects, areas of expertise, and technologies I'm familiar with. It also provides information about my preferred pronouns and willingness to collaborate.

I plan to continue updating the README as needed to reflect changes in my projects and skills.